+++ changelog.md
@@ -17,6 +17,8 @@ This section documents changes which have been merged into develop and will form
 - read_positions_on_annotation_v2.m - new code to read chosen electrodes
 - getCartoConnectorElectrodeNaming - a helper function that stores the bizarre
                                      naming conventions of catheters in Carto3!
+- comparestructure.m allows two structures to be compared, 
+                          eg userdata from two different versions of OpenEP
 
 ### Changed
 - Surface models stored as structures rather than TriRep objects for compatibility with OpenEP